feat(Select): Move core styles and behaviour to internal `Dropdown` component
