+++ packages/core/src/_internal/dropdown/index.ts
@@ -1,0 +1,2 @@
+export { default, renderOptions } from './dropdown.component';
+export type { DropdownProps, DropdownOption } from './dropdown.props';
